Collision Detection/Wall Floor
Collision detection for player movement is done and the wall and floor
are done. I'll work on sprite animation next.
Note: I think a next big thing is going to be able to scroll across the
whole mall instead of our fixed size window.

Note: Note: A second constructor for the Sprite_Base_Class has been
added for the purposes of the floor and wall which has the option to set
the draw level, DO NOT use this one, use the other where you provide a
color instead of a draw level.

Note: Note: Note: Give your sprites names, this is partially for
collision detection.