NPS Base Class finished.
I'm fairly confident that I made an NPC base class that we can use to
get the demo working. If you have any issues with it as far as needing
something done let me know and I'll do it if you don't want to. There
are instructions on how to use the NPC base class in the Summary section
above the class. Again, as if it's unclear as I wrote this fairly late
at night.